refactor(graph-db-server): group daemon lifecycle files into daemon/lifecycle/
Post-merge of dev-manu, daemon/ held 16 immediate children (1 over the
codebase-shape limit of 15), tripped by the newly-added daemonParentWatch.ts.
Five files are daemon-lifetime concerns and group naturally:

  daemon/lifecycle/
    ├── daemonGraphLifecycle.ts
    ├── daemonLockLifecycle.ts
    ├── daemonParentWatch.ts
    ├── daemonPortLifecycle.ts
    └── daemonWatcherLifecycle.ts

daemon/ now has 12 immediate children, comfortably under budget. Cross-area
imports inside the moved files (data/, state/) were rewritten to
@vt/graph-db-server/{graph,state}/* aliases — `../../` deep relatives are
banned by the relative-import-depth gate.

Also adds an agent-facing remediation message to codebase-shape.test.ts so
future agents who trip the immediate-child limit are pointed at the
correct fix (subfolders that reflect semantic structure) rather than
ratcheting the limit upward.

Tests: npm run test green (build+tier1, webapp-unit, tier2-browser all PASS).
