+++ scripts/measure-relative-paths.report.json
@@ -1,6 +1,6 @@
 {
-  "generatedAt": "2026-05-20T13:52:01.140Z",
-  "sourceFileCount": 788,
+  "generatedAt": "2026-05-21T03:40:54.658Z",
+  "sourceFileCount": 789,
   "scope": {
     "requestedRoots": [
       "webapp/src",
